refactor(scanning): Extract ScanningAlertsClientV1 client
