+++ sdcclient/secure/scanning/__init__.py
@@ -1,0 +1,3 @@
+from ._alerts import ScanningAlertsClientV1
+
+__all__ = ["ScanningAlertsClientV1"]
